+++ euclidian_distance.h
@@ -1,7 +1,7 @@
 #ifndef EUCLIDIAN_DISTANCE_H
 #define EUCLIDIAN_DISTANCE_H
 
-long double euclidian_distance(long double* pontoA, long double* pontoB, int v_size);
+double euclidian_distance(double* pontoA, double* pontoB, int v_size);
 
 
 #endif
